refactor(pi-codex-conversion): own transport recovery
